Add VPopMail

+++ bin/qmail-aio/init.sh
@@ -7,24 +7,24 @@ RESUME=$(mktemp /tmp/sqmail.XXXXXX)
 #########################
 
 #https://en.wikibooks.org/wiki/Bash_Shell_Scripting/Whiptail
-whiptail --title "Semhoun's SQMail" --msgbox "Welcome in the SQMail first time configuration\MariaDB database must already been created" 8 78
+whiptail --title "Semhoun's SQMail" --msgbox "Welcome in the SQMail first time configuration\MYSQL database must already been created" 8 78
 
-# MariaDB
-MARIADB_HOST=$(whiptail --inputbox "MariaDB Host" 8 39 "" --title "MariaDB configuration" 3>&1 1>&2 2>&3)
+# MYSQL
+MYSQL_HOST=$(whiptail --inputbox "MYSQL Host" 8 39 "" --title "MYSQL configuration" 3>&1 1>&2 2>&3)
 if [ $? != 0 ]; then echo "You canceled the script"; exit 0; fi
-MARIADB_DB=$(whiptail --inputbox "MariaDB Database" 8 39 "" --title "MariaDB configuration" 3>&1 1>&2 2>&3)
+MYSQL_DB=$(whiptail --inputbox "MYSQL Database" 8 39 "" --title "MYSQL configuration" 3>&1 1>&2 2>&3)
 if [ $? != 0 ]; then echo "You canceled the script"; exit 0; fi
-MARIADB_USER=$(whiptail --inputbox "MariaDB Username" 8 39 "" --title "MariaDB configuration" 3>&1 1>&2 2>&3)
+MYSQL_USER=$(whiptail --inputbox "MYSQL Username" 8 39 "" --title "MYSQL configuration" 3>&1 1>&2 2>&3)
 if [ $? != 0 ]; then echo "You canceled the script"; exit 0; fi
-MARIADB_PASS=$(whiptail --inputbox "MariaDB Password" 8 39 "" --title "MariaDB configuration" 3>&1 1>&2 2>&3)
+MYSQL_PASS=$(whiptail --inputbox "MYSQL Password" 8 39 "" --title "MYSQL configuration" 3>&1 1>&2 2>&3)
 if [ $? != 0 ]; then echo "You canceled the script"; exit 0; fi
 
 cat >> "${RESUME}" <<EOF
 Mysql configration will be:
-  - host : ${MARIADB_HOST}
-  - database : ${MARIADB_DB}
-  - username : ${MARIADB_USER}
-  - password : ${MARIADB_PASS}
+  - host : ${MYSQL_HOST}
+  - database : ${MYSQL_DB}
+  - username : ${MYSQL_USER}
+  - password : ${MYSQL_PASS}
   
 EOF
 
@@ -75,8 +75,19 @@ Domain configuration will be:
   
 EOF
 
+ROUNDCUBE_SUPPORT=$(whiptail --inputbox "Roundcube support url (with https:// or mailto;" 12 50 "mailto:help@${DEFAULT_DOMAIN}" --title "Roundcube Webmail Configuration" 3>&1 1>&2 2>&3)
+if [ $? != 0 ]; then echo "You canceled the script"; exit 0; fi
+ROUNDCUBE_NAME=$(whiptail --inputbox "Roundcube name" 8 39 "Roundcube Webmail" --title "RoundCube Webmail Configuration" 3>&1 1>&2 2>&3)
+if [ $? != 0 ]; then echo "You canceled the script"; exit 0; fi
+cat >> "${RESUME}" << EOF
+Roundcube Webmail  configuration will be:
+  - support url : ${ROUNDCUBE_SUPPORT}
+  - roundcube name : ${ROUNDCUBE_NAME}
+  
+EOF
+
 # Resume
-whiptail --textbox "${RESUME}" 25 78
+whiptail --textbox "${RESUME}" 30 78
 rm "${RESUME}"
 
 if !(whiptail --title "Set configuration" --yesno "Apply the configuration." 8 78); then
@@ -85,9 +96,14 @@ if !(whiptail --title "Set configuration" --yesno "Apply the configuration." 8 7
 fi
 
 #-----------------------------------------------------------------------------#
+
 #########################
 # Create config
 #########################
+export MYSQL_USER=${MYSQL_USER}
+export MYSQL_PASS=${MYSQL_PASS}
+export MYSQL_DB=${MYSQL_DB}
+export MYSQL_HOST=${MYSQL_HOST}
 
 # Default config
 echo "MAILER-DAEMON" > /var/qmail/control/bouncefrom
@@ -100,17 +116,17 @@ openssl dhparam -out /ssl/dovecot-dhparam 2048
 
 # Creation configuration
 cat > /var/qmail/control/mysql.conf << EOF
-MYSQL_USER=${MARIADB_USER}
-MYSQL_PASS=${MARIADB_PASS}
-MYSQL_DB=${MARIADB_DB}
-MYSQL_HOST=${MARIADB_HOST}
+MYSQL_USER=${MYSQL_USER}
+MYSQL_PASS=${MYSQL_PASS}
+MYSQL_DB=${MYSQL_DB}
+MYSQL_HOST=${MYSQL_HOST}
 EOF
-echo "${MARIADB_HOST}|0|${MARIADB_USER}|${MARIADB_PASS}|${MARIADB_DB}" > /var/vpopmail/etc/vpopmail.mysql
+echo "${MYSQL_HOST}|0|${MYSQL_USER}|${MYSQL_PASS}|${MYSQL_DB}" > /var/vpopmail/etc/vpopmail.mysql
 cat > /var/qmail/control/spamassassin_sql.cf << EOF
 # User prefs
-user_scores_dsn DBI:mysql:${MARIADB_DB}:${MARIADB_HOST}
-user_scores_sql_username ${MARIADB_USER}
-user_scores_sql_password ${MARIADB_PASS}
+user_scores_dsn DBI:mysql:${MYSQL_DB}:${MYSQL_HOST}
+user_scores_sql_username ${MYSQL_USER}
+user_scores_sql_password ${MYSQL_PASS}
 user_scores_sql_custom_query     SELECT preference, value FROM spam_prefs WHERE username = _USERNAME_ OR username = '\$GLOBAL' OR username = CONCAT('%',_DOMAIN_) ORDER BY username ASC
 EOF
 echo -n "${CONCURRENCY_INCOMING}" > /var/qmail/control/concurrencyincoming
@@ -121,53 +137,7 @@ echo "${QUEUELIFETIME}" > /var/qmail/control/queuelifetime
 
 # VPopmail configuration
 echo "${DEFAULT_DOMAIN}" > /var/vpopmail/etc/defaultdomain
-cat > /var/vpopmail/etc/vlimits.default << 'EOF'
-# Default limits file.  This file is used for domains without a
-# .qmailadmin-limits file.
-
-# maximums for each account type, -1 = unlimited
-maxpopaccounts          -1
-maxforwards             -1
-maxautoresponders       -1
-maxmailinglists         -1
-
-# quota for entire domain, in megabytes
-# example shows a domain with a 100MB quota and a limit of 10,000 messages
-#quota                  100
-#maxmsgcount            10000
-
-# default quota for newly created users (in bytes)
-# example shows a user with a 10MB quota and a limit of 1000 messages
-#default_quota          10485760
-#default_maxmsgcount    1000
-
-# uncomment the following lines to disable certain features
-#disable_pop
-#disable_imap
-#disable_dialup
-#disable_password_changing
-#disable_external_relay
-#disable_smtp
-#disable_webmail
-#disable_spamassassin
-#delete_spam
-#disable_maildrop
-
-
-# Set bitflags on account management for non-postmaster admins.
-# To disable certain features, add the following bits:
-#   Create = 1, Modify = 2, Delete = 4
-# So, to allow modification but not creation or deletion of
-# POP/IMAP accounts, set perm_account to 5.
-
-perm_account            0
-perm_alias              0
-perm_forward            0
-perm_autoresponder      0
-perm_maillist           0
-perm_quota              0
-perm_defaultquota       0
-EOF
+cp /qmail-aio/templates/vlimits.default /var/vpopmail/etc/vlimits.default
 
 # Qmail configuration from /package/mail/sqmail/sqmail/src/config-fast.sh
 echo "${SMTP_SERVER}" > /var/qmail/control/me
@@ -185,41 +155,9 @@ chown alias.sqmail .qmail*
 chmod 644 .qmail*
  
 # Dovecot
-cat > /var/qmail/control/dovecot-sql.conf.ext << EOF
-driver = mysql
-connect = host=${MARIADB_HOST} dbname=${MARIADB_DB} user=${MARIADB_USER} password=${MARIADB_PASS}
-default_pass_scheme = MD5-CRYPT
-
-#  USER LIMITS via vpopmail.pw_gid filed was currently removed
-password_query = \
-	SELECT \
-		CONCAT(vpopmail.pw_name, '@', vpopmail.pw_domain) AS user, \
-  		vpopmail.pw_passwd AS password, \
-                vpopmail.pw_dir AS userdb_home, \
-                89 AS userdb_uid, \
-                89 AS userdb_gid, \
-                CONCAT('*:bytes=', REPLACE(SUBSTRING_INDEX(vpopmail.pw_shell, 'S', 1), 'NOQUOTA', '0')) AS userdb_quota_rule \
-	FROM vpopmail \
-		LEFT JOIN aliasdomains ON aliasdomains.alias='%d' \
-		LEFT JOIN limits ON limits.domain = '%d' \
-	WHERE \
-		vpopmail.pw_name='%n' \
-		AND \
-		(vpopmail.pw_domain='%d' OR vpopmail.pw_domain=aliasdomains.domain)
-
-user_query = \
-	SELECT \
-		vpopmail.pw_dir AS home, \
-	  	89 AS uid, \
-  		89 AS gid \
-  	FROM vpopmail \
-  	WHERE \
-  		vpopmail.pw_name='%n' \
-		AND \
-		vpopmail.pw_domain='%d'
-
-iterate_query = SELECT CONCAT(pw_name,'@',pw_domain) AS user FROM vpopmail
-EOF
+cat /qmail-aio/templates/dovecot-sql.conf.ext | envsubst \
+		'$MYSQL_USER $MYSQL_PASS $MYSQL_HOST $MYSQL_DB' \
+		> /var/qmail/control/dovecot-sql.conf.ext
 chown root.root /var/qmail/control/dovecot-sql.conf.ext
 chmod 600 /var/qmail/control/dovecot-sql.conf.ext
 
@@ -241,26 +179,7 @@ chown -R vpopmail.vchkpw /var/vpopmail/domains
 /var/vpopmail/bin/vadddomain ${DEFAULT_DOMAIN} "${POSTMASTER_PWD}"
 
 # SpamAssassin DB
-echo "CREATE TABLE spam_prefs (
-  id int(8) UNSIGNED NOT NULL,
-  username varchar(128) NOT NULL DEFAULT '',
-  preference varchar(64) NOT NULL DEFAULT '',
-  value varchar(128) DEFAULT NULL,
-  added datetime NOT NULL DEFAULT current_timestamp(),
-  modified timestamp NOT NULL DEFAULT current_timestamp() ON UPDATE current_timestamp(),
-  UNIQUE KEY id (id),
-  KEY preference (preference),
-  KEY username (username)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1 COMMENT='Spamassassin Preferences';
-INSERT INTO spam_prefs (id, username, preference, value, added, modified) VALUES
-(1, '\$GLOBAL', 'required_hits', '7.0', '2003-10-11 00:00:00', '2013-09-10 00:00:00'),
-(2, '\$GLOBAL', 'report_safe', '0', '2003-10-11 00:00:00', '2013-09-10 00:00:00'),
-(3, '\$GLOBAL', 'fold_headers', '1', '2003-10-11 00:00:00', '2013-09-10 00:00:00'),
-(4, '\$GLOBAL', 'add_header', 'all Level _STARS(*)_', '2003-10-11 00:00:00', '2013-09-10 00:00:000'),
-(5, '\$GLOBAL', 'rewrite_header', 'Subject [SPAM]', '2003-10-11 00:00:00', '2013-09-10 00:00:00'),
-(6, '\$GLOBAL', 'ok_languages', 'en fr', '2003-10-11 00:00:00', '2013-09-10 00:00:00'),
-(7, '\$GLOBAL', 'refuse_threshold', '9', '2003-10-11 00:00:00', '2013-09-10 00:00:00');
-" | mysql -h ${MARIADB_HOST} -u ${MARIADB_USER} -p"${MARIADB_PASS}" ${MARIADB_DB}
+cat /qmail-aio/templates/spamassassin.sql | mysql -h ${MYSQL_HOST} -u ${MYSQL_USER} -p"${MYSQL_PASS}" ${MYSQL_DB}
 
 # Rules cdb
 cat > /var/qmail/control/rules.smtpsub << EOF
@@ -276,6 +195,20 @@ done
 echo ":allow,QHPSI='clamdscan',QHPSIARG1='--no-summary',MFDNSCHECK='',BADMIMETYPE='',BADLOADERTYPE='M',HELOCHECK='.',TARPITCOUNT='5',TARPITDELAY='20',QMAILQUEUE='bin/qmail-queuescan'" >> /var/qmail/control/rules.smtpd
 echo ":allow,QHPSI='clamdscan',QHPSIARG1='--no-summary',MFDNSCHECK='',BADMIMETYPE='',BADLOADERTYPE='M',HELOCHECK='.',TARPITCOUNT='5',TARPITDELAY='20',QMAILQUEUE='bin/qmail-queuescan'" >> /var/qmail/control/rules.smtpsd
 /usr/local/bin/qmailctl cdb
+
+# Generate roundcube config
+cat > /var/qmail/control/roundcube.conf << EOF
+export MYSQL_USER=${MYSQL_USER}
+export MYSQL_PASS=${MYSQL_PASS}
+export MYSQL_DB=${MYSQL_DB}
+export MYSQL_HOST=${MYSQL_HOST}
+export DES_KEY=`apg -MSNCL -m 24 -x 24 -n 1`
+export SUPPORT_URL="${ROUNDCUBE_SUPPORT}"
+export PRODUCT_NAME="${ROUNDCUBE_NAME}"
+EOF
+
+# Roundcube DB
+cat /qmail-aio/templates/roundcube.sql | mysql -h ${MYSQL_HOST} -u ${MYSQL_USER} -p"${MYSQL_PASS}" ${MYSQL_DB}
 
 echo "============================"
 echo " QMail AllInOne initialized"
